% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creole_lines.R
\name{creole_lines}
\alias{creole_lines}
\title{creole_lines}
\usage{
creole_lines(outDir, origin = NA, subFrac = 0.6, numPts = 10000,
  numIters = 100, GoI = NA)
}
\arguments{
\item{origin}{integer value corresponding to the origin cluster of the data. Defaults to putative origin identified in creole_map}

\item{subFrac}{sub-fraction of cells selected at each iteration of expression trend estimation, given as percent. Defaults to 0.66}

\item{numPts}{number of points or resolution of final output trend estimation. Defaults to 10000}

\item{numIters}{number of estimation iterations. Defaults to 100}

\item{GoI}{List of Genes of Interest. 'NA' results in all genes, which could significantly increase processing time}
}
\description{
This function calculates consensus trends of gene expression through each lineage established by creole_map
}
\examples{
creole_lines()
}
\keyword{creole_lines}

