% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creole_map.R
\name{creole_map}
\alias{creole_map}
\title{creole_map}
\usage{
creole_map(data, is.scaled = FALSE, outDir, numClusters, numDims, sigma,
  doplot3d = FALSE, dimD = "PCA")
}
\arguments{
\item{data}{scaled data as an n x G data frame object}

\item{is.scaled}{logical input to designate input data as scaled or not (z-transform)}

\item{outDir}{output directory}

\item{numClusters}{number of clusters to use for lineage calculation}

\item{numDims}{number of dimensions to use in reducing data dimensionality}

\item{sigma}{gaussian kernal width for diffusion calculation}

\item{doplot3d}{logical whether to launch a window of the 3-D subspace. Defaults to FALSE}

\item{dimD}{method of dimensionality reduction, currently PCA and Diffusion Mapping are implemented}
}
\description{
This function maps input data and calculates associated lineages
}
\examples{
creole_map()
}
\keyword{creole_map}

