% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creole_plots.R
\name{creole_plots}
\alias{creole_plots}
\title{creole_plots}
\usage{
creole_plots(outDir, figW = 100, figH = 60, figU = "mm", figF = "eps",
  figD = 350)
}
\arguments{
\item{figW}{desired figure width}

\item{figH}{desired figure height}

\item{figU}{figure measurement units}

\item{figF}{desired figure format}

\item{figD}{figure DPI}

\item{origin}{integer}
}
\description{
This function calculates consensus trends of gene expression through each lineage established by creole_map
}
\examples{
creole_plot()
}
\keyword{creole_plot}

