% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffMap.R
\name{DiffMap}
\alias{DiffMap}
\title{DiffMap}
\usage{
DiffMap(data, sigma)
}
\arguments{
\item{data}{scaled data as an n x G data frame object}

\item{sigma}{kernel width}
}
\description{
DiffMap returns the diffusion map corrdinates calculated from input data and sigma, the gaussian kernel width
}
\keyword{Diffusion}
\keyword{Map}

