% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLineages.R
\name{GetLineages}
\alias{GetLineages}
\title{GetLineages}
\usage{
GetLineages(dmst, origin.cluster)
}
\arguments{
\item{dmst}{Minimum Spanning Tree of cluster centroids objec}

\item{origin.cluster}{Cluster of origin}
}
\description{
This function identifies the lineages from a given MST and cluster of origin. Terminal nodes are defines as all nodes not the origin with a singular edge. getLineages returns all possible lineages beginning at the origin
}
\examples{
GetLineages()
}
\keyword{getLineages}

