% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeMST.R
\name{makeMST}
\alias{makeMST}
\title{makeMST}
\usage{
makeMST(centers)
}
\arguments{
\item{centers}{n x d matrix or dataframe object of n cluster centers in d dimensions}
}
\description{
makeMST calculates a minimum spanning tree of cluster centroids
}
\keyword{makeMST}

