% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSyntheticData.R
\name{makeSyntheticData}
\alias{makeSyntheticData}
\title{makeSyntheticData}
\usage{
makeSyntheticData()
}
\description{
This function generates a synthetic dataset for use in example documentation
}
\examples{
makeSyntheticData()
}
\keyword{makeSyntheticData}

