% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDM3D.R
\name{plotDM3D}
\alias{plotDM3D}
\title{plotDM3D}
\usage{
plotDM3D(init.DM, KM, dmst, fig.dir)
}
\arguments{
\item{KM}{K-means clustering object}

\item{dmst}{Minimum Spanning Tree object}

\item{fig.dir}{Figure directory}

\item{DM}{Diffusion map coordinates}
}
\description{
plotDM3D returns opens a new window and a 3-dimensional figure viewer to better visualize the diffusion map and lineage mapping
}
\keyword{plotDM3D}

