% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaOpt.R
\name{sigmaOpt}
\alias{sigmaOpt}
\title{sigmaOpt}
\usage{
sigmaOpt(data, minSigma, numSteps, stepSize, is.scaled = FALSE)
}
\arguments{
\item{data}{scaled data as an n x G data frame object}

\item{minSigma}{lower bound for search space}

\item{numSteps}{number of steps to take in optimizing kernal}

\item{stepSize}{incremental width of search space}

\item{is.scaled}{logical value indicates whether data is scaled}
}
\description{
sigmaOpt is a course optimization routine for gaussian kernal width selection for use in calculating diffusion map coordinates
}
\examples{
sigmaOpt()
}
\keyword{sigmaOpt}

