% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VbSCO.R
\name{VbSCO}
\alias{VbSCO}
\title{VbSCO}
\usage{
VbSCO(DM, k, c)
}
\arguments{
\item{DM}{Diffusion Map coordinates}

\item{k}{K-means cluster array}

\item{c}{list of cluster that define the lineage}
}
\description{
The Vector-based Single Cell Ordering (VbSCO) algorithm is designed to project cells of lineage-defining clusters onto centroid-binding vectors
}
\examples{
VbSCO()
}
\keyword{VbSCO}

